refactor(fuse-backend): split the crate into focused modules (#71)
lib.rs was a single ~700-line file. Decompose by responsibility:

- convert.rs — KisekiFS <-> libc/ABI conversions (attr_to_stat64, errno/type maps)
- fs.rs      — the FileSystem trait impl (KisekiFsBackend)
- session.rs — mount + worker-thread pool (FbrSession, mount, serve)
- lib.rs     — thin root: crate docs, module wiring, public re-exports

No behaviour change; mounted smoke still passes.

Refs #71
